[feat] 공통 레이아웃을 _app.tsx로 분리함
